requirements.txt
gymnasium
numpy
scipy
torch
tensorboard
pyyaml
pandas
pytest
matplotlib
ruff
black
pre-commit
